[dev]: Fix the pfe-navigation mobile login/language links
