+++ elements/pfe-navigation/demo/index.html
@@ -74,7 +74,7 @@
     <form slot="mobile-search" hidden>
       <input type="text" name="search" value="" placeholder="Enter your search term"
         style="height: 30px; width: 60%; margin-right: 10px;">
-      <pfe-cta priority="primary"><a href="#">Search</a></pfe-cta>
+      <pfe-cta priority="primary"><a href="http://www.google.com">Search</a></pfe-cta>
     </form>
     <!-- Main navigation -->
     <!-- <h2 slot="mobile-menu--label" hidden>Menu</h2> -->
@@ -495,27 +495,27 @@
     </pfe-navigation-main>
     <!-- Language -->
     <pfe-navigation-item slot="language" pfe-icon="web-globe">
-      <h2 slot="trigger"><a href="/url-to-language-page">English</a></h2>
+      <h2 slot="trigger"><a href="http://www.foo.com">English</a></h2>
       <div slot="tray" hidden>
         <div class="pfe-navigation-item__tray--container">
           <p>Language switcher</p>
         </div>
       </div>
     </pfe-navigation-item>
-    <a href="/url-to-language-page" slot="mobile-language" pfe-icon="web-globe">English</a>
+    <a href="http://www.foo.com" slot="mobile-language" pfe-icon="web-globe">English</a>
     <!-- Log in -->
     <pfe-navigation-item slot="login" pfe-icon="web-user">
-      <h2 slot="trigger"><a href="/login">Log in</a></h2>
+      <h2 slot="trigger"><a href="http://www.bar.com">Log in</a></h2>
       <div slot="tray" hidden>
         <div class="pfe-navigation-item__tray--container">
           <p>Log in things</p>
         </div>
       </div>
     </pfe-navigation-item>
-    <a href="/login" slot="mobile-login" pfe-icon="web-user">Login/Register</a>
+    <a href="http://www.bar.com" slot="mobile-login" pfe-icon="web-user">Login/Register</a>
     <!-- Site switcher -->
     <pfe-navigation-item slot="site-switcher" pfe-icon="web-grid-3x3">
-      <h2 slot="trigger"><a href="/url-to-rh-websites-page">Websites</a></h2>
+      <h2 slot="trigger"><a href="http://www.bat.com">Websites</a></h2>
       <div slot="tray" hidden>
         <div class="pfe-navigation-item__tray--container">
           <h3>Site switcher things</h3>
