+++ examples/index.html
@@ -29,6 +29,7 @@
 					<li><a href="../elements/pfe-number/demo">pfe-number</a></li>
 					<li><a href="../elements/pfe-page-status/demo">pfe-page-status</a></li>
 					<li><a href="../elements/pfe-progress-indicator/demo">pfe-progress-indicator</a></li>
+					<li><a href="../elements/pfe-styles/demo">pfe-styles</a></li>
 					<li><a href="../elements/pfe-tabs/demo">pfe-tabs</a></li>
 					<li><a href="../elements/pfe-toast/demo">pfe-toast</a></li>
 				</ul>
